app/pyproject.toml
[tool.poetry]
name = "notification-service"
version = "0.1.0"
description = ""
authors = ["Elnar Mengelbaev <elnar_men@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
environs = "^9.5.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
celery = {extras = ["redis"], version = "^5.3.5"}
flower = "^2.0.1"
requests = "^2.31.0"
drf-yasg = "^1.21.7"
factory-boy = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
